requirements.txt
langchain
langchain-google-genai
langchain-groq
google-generativeai
groq
python-dotenv
requests